- 修复辉光魔力池在 JEI 的错误颜色显示。 - 修复 ImmersiveEngineering 的结构检查代码修复不起作用的问题。 - 新增 ImmersiveEngineering 的油罐修复（实际上是限制一部分合成）。 - 新增 IC2 能量网络优化。 - 加大 AstralSorcery 的 PlayerProgress 的修改力度，防止边界情况崩溃。 - 新增 AstralSorcery 的 PlayerAttributeMap，防止边界情况崩溃。 - 尝试移除几个模组的强依赖。

+++ src/main/java/github/kasuminova/novaeng/common/hypernet/NetNodeImpl.java
@@ -61,7 +61,7 @@ public class NetNodeImpl extends NetNode {
 
         int currentParallelism = event.getActiveRecipe().getParallelism();
         if (currentParallelism > 1) {
-            int max = (int) Math.min(currentParallelism, (double) (generation / pointRequired));
+            int max = (int) Math.min(currentParallelism, generation / pointRequired);
             event.setParallelism(max);
         }
     }
